seperate shared examples with specs remove code which was causing `WARNING`s add `FIXME`s and some code optimization indent the code
