+++ lib/cms/concerns/categorizable.rb
@@ -2,6 +2,6 @@ module Categorizable
   extend ActiveSupport::Concern
 
   included do
-    field :categories,  type: Array
+    field :categories, type: Array
   end
 end
