+++ spec/concerns/renderable_spec.rb
@@ -1,9 +1,5 @@
 require 'spec_helper'
 
-class RenderableModel < Cms::Models::Document
-  include Renderable
-end
-
 describe Renderable do
   pending 'todo'
 end
