+++ spec/models/block_spec.rb
@@ -1,19 +1,13 @@
 require 'spec_helper'
-require 'concerns/document_spec'
-require 'concerns/content_item_spec'
-require 'concerns/listable_spec'
-require 'concerns/extendable_spec'
-require 'concerns/images_spec'
-require 'concerns/templates_spec'
 
 RSpec.describe Cms::Models::Block, type: :model do
-  
+
   let(:block) {FactoryGirl.create(:block)}
-  
+
   it_behaves_like "Documentable" do
     let(:document_model) {FactoryGirl.create(:block)}
   end
-  
+
   it_behaves_like "ContentItem" do
     let(:name) { Faker::Lorem.characters(10) }
     let(:title) { expected_title_translations['en'] }
@@ -22,26 +16,26 @@ RSpec.describe Cms::Models::Block, type: :model do
     let(:description) { expected_description_translations['en'] }
     let(:content_item_model) {FactoryGirl.create(:block, name: name, title: title, description: description, categories: categories, tags: tags)}
   end
-  
+
   it_behaves_like "Blueprintable" do
     let(:prototype) { FactoryGirl.create(:block) }
     let(:blueprint) { prototype.blueprints.create name: Faker::Lorem.characters(10)  }
   end
-  
+
   it_behaves_like "Listable" do
     let(:listable_model) {FactoryGirl.create(:block)}
     let(:listable_model1) { FactoryGirl.create(:block, position: 2 ) }
     let(:listable_model2) { FactoryGirl.create(:block, position: 3 ) }
   end
-  
+
   it_behaves_like "Extendable" do
     let(:extendable_model) {FactoryGirl.create(:block, generic_attributes_attributes: generic_attributes_array)}
   end
-  
+
   it_behaves_like "Images" do
     let(:image_model) {FactoryGirl.create(:block) }
   end
-  
+
   it_behaves_like "Templates" do
     let(:template_model) {FactoryGirl.create(:block) }
   end
@@ -64,7 +58,7 @@ RSpec.describe Cms::Models::Block, type: :model do
     block.content = nil
     expect(block).to be_invalid
   end
-  
+
 
   # it "should have trackable history and counts" do
   #   user = FactoryGirl.create(:user)
