+++ spec/support/factory_girl.rb
@@ -1,0 +1,5 @@
+RSpec.configure do |config|
+  config.include FactoryGirl::Syntax::Methods
+  FactoryGirl.definition_file_paths = %w{./spec/factories}
+  FactoryGirl.find_definitions
+end
