+++ spec/support/shared_examples/renderable.rb
@@ -1,0 +1,3 @@
+class RenderableModel < Cms::Models::Document
+  include Renderable
+end
