+++ spec/support/vcr.rb
@@ -1,0 +1,14 @@
+require 'vcr'
+require 'webmock'
+require 'webmock/rspec'
+
+RSpec.configure do |config|
+  WebMock.disable_net_connect!(allow_localhost: true)
+end
+
+VCR.configure do |c|
+  #the directory where your cassettes will be saved
+  c.cassette_library_dir = 'spec/vcr'
+  # your HTTP request service.
+  c.hook_into :webmock
+end
